Added Bae chart module to the project
